# PERF: op(frame, series) when series is not EA