--- a/pandas/core/ops/__init__.py
+++ b/pandas/core/ops/__init__.py
@@ -340,20 +340,18 @@ def dispatch_to_series(left, right, func, str_rep=None, axis=None):
             right = np.asarray(right)
 
             def column_op(a, b):
-                return {i: func(a._ixs(i, axis=1), b[i]) for i in range(len(a.columns))}
+                return {i: func(a.iloc[:, i], b[i]) for i in range(len(a.columns))}
 
         else:
 
             def column_op(a, b):
-                return {
-                    i: func(a._ixs(i, axis=1), b.iloc[i]) for i in range(len(a.columns))
-                }
+                return {i: func(a.iloc[:, i], b.iloc[i]) for i in range(len(a.columns))}
 
     elif isinstance(right, ABCSeries):
         assert right.index.equals(left.index)  # Handle other cases later
 
         def column_op(a, b):
-            return {i: func(a._get_column_array(i), b) for i in range(len(a.columns))}
+            return {i: func(a.iloc[:, i], b) for i in range(len(a.columns))}
 
     else:
         # Remaining cases have less-obvious dispatch rules