We have the same optimization in place for the flex ops.

ipython results for the asv this adds:

```
In [3]: arr = np.arange(10 ** 6).reshape(100, -1)                                                                                                                                                                   
In [4]: df = pd.DataFrame(arr)                                                                                                                                                                                         
In [5]: df["C"] = 1.0                                                                                                                                                                                               
In [6]: row = df.iloc[0]                                                                                                                                                                                            

In [11]: %timeit df + row                                                                                                                                                                                           
1.92 s ± 71.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)   # <-- master
1 ms ± 19.2 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <-- PR
```