@jbrockmendel 2020-04-19
hmm when i run  `asv dev` locally it stops at 50% but with exit code 0.  any guesses @datapythonista ?