@datapythonista 2020-04-21
> hmm when i run `asv dev` locally it stops at 50% but with exit code 0. any guesses @datapythonista ?

I tried `asv run` locally with master, and it worked well until the end. Is it just `asv dev`, or just this PR that is stopping half way? I don't see why this should happen.