@datapythonista 2020-04-21
Tried `asv dev` locally, and also stopped at 50%. Is there an issue for it? I see that last release of `asv` was a year ago, so I guess it's something in our side.