@jbrockmendel 2020-04-21
Looks like the asv issue is actually that this causes an unrelated benchmark to fail.  Am troubleshooting now and will add an appropriate test.