@jbrockmendel 2020-04-22
> lgtm. can you add a whatsnew note & an asv that covers this.

done